backend/requirements.txt
fastapi
uvicorn
redis
pydantic
sqlmodel
asyncpg
python-dotenv
python-socketio[asyncio_client]
aiofiles
celery
